Moved code from custom_partner_fields to school_management + deleted custom_partner_fields_module
