--- custom_partner_fields/README.rst
@@ -1,55 +1,0 @@
-=====================
-Custom Partner Fields
-=====================
-
-.. 
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! This file is generated by oca-gen-addon-readme !!
-   !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:6483b7d898c851d351b880f22b89f1fcb78dd565f8327f69d15fe3ecb0283f9a
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
-    :target: https://odoo-community.org/page/development-status
-    :alt: Beta
-.. |badge2| image:: https://img.shields.io/badge/licence-LGPL--3-blue.png
-    :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
-    :alt: License: LGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-ito--invest--lu%2Fhorizon-lightgray.png?logo=github
-    :target: https://github.com/ito-invest-lu/horizon/tree/16.0/custom_partner_fields
-    :alt: ito-invest-lu/horizon
-
-|badge1| |badge2| |badge3|
-
-This module...
-
-**Table of contents**
-
-.. contents::
-   :local:
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/ito-invest-lu/horizon/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/ito-invest-lu/horizon/issues/new?body=module:%20custom_partner_fields%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Do not contact contributors directly about support or help with technical issues.
-
-Credits
-=======
-
-Authors
-~~~~~~~
-
-* Deuse
-
-Maintainers
-~~~~~~~~~~~
-
-This module is part of the `ito-invest-lu/horizon <https://github.com/ito-invest-lu/horizon/tree/16.0/custom_partner_fields>`_ project on GitHub.
-
-You are welcome to contribute.
