--- custom_partner_fields/__init__.py
@@ -1,1 +1,0 @@
-from . import models
