--- custom_partner_fields/models/__init__.py
@@ -1,1 +1,0 @@
-from . import res_partner
