--- custom_partner_fields/readme/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-This module...
