--- custom_partner_fields/security/ir.model.access.csv
@@ -1,5 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-custom_partner_fields_erasmus,school_program_erasmus,model_custom_partner_fields_erasmus,school_management.group_employee,1,1,1,1
-custom_partner_fields_internship,school_program_internship,model_custom_partner_fields_internship,school_management.group_employee,1,1,1,1
-custom_partner_fields_memoir_title,school_program_memoir_title,model_custom_partner_fields_memoir_title,school_management.group_employee,1,1,1,1
-custom_partner_fields_access_title,school_program_access_title,model_custom_partner_fields_access_title,school_management.group_employee,1,1,1,1
