+++ school_management/security/ir.model.access.csv
@@ -60,3 +60,7 @@ individual_bloc.tagteacher,model_individual_bloc.tagbase.group_teacher,model_sch
 individual_bloc.tagbase.group_public,model_individual_bloc.tagbase.group_public,model_school_individual_bloc_tag,base.group_public,1,0,0,0
 individual_bloc.tagbase.group_portal,model_individual_bloc.tagbase.group_portal,model_school_individual_bloc_tag,base.group_portal,1,0,0,0
 school_program_student,school_program_student,model_school_program,group_student,1,0,0,0
+custom_partner_fields_access_title,school_program_access_title,model_custom_partner_fields_access_title,school_management.group_employee,1,1,1,1
+custom_partner_fields_erasmus,school_program_erasmus,model_custom_partner_fields_erasmus,school_management.group_employee,1,1,1,1
+custom_partner_fields_internship,school_program_internship,model_custom_partner_fields_internship,school_management.group_employee,1,1,1,1
+custom_partner_fields_memoir_title,school_program_memoir_title,model_custom_partner_fields_memoir_title,school_management.group_employee,1,1,1,1
