+++ school_management/views/res_partner_view.xml
@@ -758,5 +758,75 @@
         sequence="99"
     />
 
+    <!-- custom_partner_fields -->
+    <record id="view_partners_custom_fields" model="ir.ui.view">
+        <field name="name">view.res.partner.form.school</field>
+        <field name="model">res.partner</field>
+        <field name="inherit_id" ref="base.view_partner_form" />
+        <field name="arch" type="xml">
+            <xpath expr="//h1//field[@name='name']/.." position="after">
+                <field name="second_first_name" string="Autre(s) prénom(s)" />
+            </xpath>
+        </field>
+    </record>
+
+    <record id="view_student_form_school_pages" model="ir.ui.view">
+        <field name="name">view.res.partner.form.school.pages</field>
+        <field name="model">res.partner</field>
+        <field name="inherit_id" ref="school_management.view_partners_form_school" />
+        <field name="arch" type="xml">
+            <xpath expr="//notebook[last()]" position="inside">
+                <page name="access_titles_page" string="Titre d'accès">
+                    <group>
+                        <field name="admission_exam_date" />
+                        <field name="access_titles_ids" mode="tree">
+                            <tree>
+                                <field name="name" />
+                                <field name="title" />
+                                <field name="establishment" />
+                                <field name="city" />
+                                <field name="country_id" />
+                                <field name="date" />
+                                <field name="type" />
+                                <field name="country" />
+                            </tree>
+                        </field>
+                    </group>
+                </page>
+
+                <page name="erasmus_page" string="Divers">
+                    <group>
+                        <field name="memoir_titles_ids">
+                            <tree>
+                                <field name="name" />
+                            </tree>
+                        </field>
+                        <field name="internships_ids" mode="tree">
+                            <tree>
+                                <field name="establishment" />
+                                <field name="city" />
+                                <field name="country_id" />
+                                <field name="start_date" />
+                                <field name="end_date" />
+                                <field name="country" />
+                            </tree>
+                        </field>
+                        <field name="erasmus_ids" mode="tree">
+                            <tree>
+                                <field name="establishment" />
+                                <field name="city" />
+                                <field name="country_id" />
+                                <field name="start_date" />
+                                <field name="end_date" />
+                                <field name="language" />
+                                <field name="country" />
+                            </tree>
+                        </field>
+                    </group>
+                </page>
+            </xpath>
+        </field>
+    </record>
+
 
 </odoo>
